requirements.txt
streamlit
streamlit-option-menu
numpy
pandas
pillow
opencv-python-headless
matplotlib
scikit-learn
tensorflow
psycopg2-binary
openpyxl
